refactor(api): move `exit`, `relaunch` APIs to `process` module
