+++ .changes/process-api.md
@@ -1,0 +1,5 @@
+---
+"api": patch
+---
+
+Move `exit` and `relaunch` APIs from `app` to `process` module.
