+++ core/tauri/src/endpoints/app.rs
@@ -2,10 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-use std::process::exit;
-
 use super::InvokeResponse;
-use crate::api::{process::restart, PackageInfo};
+use crate::api::PackageInfo;
 use serde::Deserialize;
 
 /// The API descriptor.
@@ -18,11 +16,6 @@ pub enum Cmd {
   GetAppName,
   /// Get Tauri Version
   GetTauriVersion,
-  /// Relaunch application
-  Relaunch,
-  /// Close application with provided exit_code
-  #[serde(rename_all = "camelCase")]
-  Exit { exit_code: i32 },
 }
 
 impl Cmd {
@@ -31,16 +24,6 @@ impl Cmd {
       Self::GetAppVersion => Ok(package_info.version.into()),
       Self::GetAppName => Ok(package_info.name.into()),
       Self::GetTauriVersion => Ok(env!("CARGO_PKG_VERSION").into()),
-      Self::Relaunch => Ok({
-        restart();
-        ().into()
-      }),
-      Self::Exit { exit_code } => {
-        // would be great if we can have a handler inside tauri
-        // who close all window and emit an event that user can catch
-        // if they want to process something before closing the app
-        exit(exit_code);
-      }
     }
   }
 }
