+++ tooling/api/rollup.config.js
@@ -27,7 +27,8 @@ export default [
       window: './src/window.ts',
       cli: './src/cli.ts',
       notification: './src/notification.ts',
-      globalShortcut: './src/globalShortcut.ts'
+      globalShortcut: './src/globalShortcut.ts',
+      process: './src/process.ts'
     },
     treeshake: true,
     perf: true,
